refactor!: reconstruct package structure
