+++ src/main/java/com/gamenawa/eeeeweeee/domain/games/dto/Game.java
@@ -1,4 +1,4 @@
-package com.gamenawa.eeeeweeee.games;
+package com.gamenawa.eeeeweeee.domain.games.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
