+++ src/main/java/com/gamenawa/eeeeweeee/domain/games/service/IGameInfoSearcher.java
@@ -1,0 +1,7 @@
+package com.gamenawa.eeeeweeee.domain.games.service;
+
+import com.gamenawa.eeeeweeee.domain.games.dto.Game;
+
+public interface IGameInfoSearcher {
+    Game getGameInfoByTitle(String title);
+}
