+++ src/main/java/com/gamenawa/eeeeweeee/domain/games/service/IGameScoreSearcher.java
@@ -1,0 +1,7 @@
+package com.gamenawa.eeeeweeee.domain.games.service;
+
+import com.gamenawa.eeeeweeee.domain.games.dto.GameScore;
+
+public interface IGameScoreSearcher {
+    GameScore getGameScoreByTitle(String title);
+}
