+++ src/main/java/com/gamenawa/eeeeweeee/domain/games/service/IGamesService.java
@@ -1,4 +1,4 @@
-package com.gamenawa.eeeeweeee.games;
+package com.gamenawa.eeeeweeee.domain.games.service;
 
 public interface IGamesService {
     String getGameJsonByTitle(String title);
