--- src/main/java/com/gamenawa/eeeeweeee/games/IGameInfoSearcher.java
@@ -1,5 +1,0 @@
-package com.gamenawa.eeeeweeee.games;
-
-public interface IGameInfoSearcher {
-    Game getGameInfoByTitle(String title);
-}
