--- src/main/java/com/gamenawa/eeeeweeee/games/IGameScoreSearcher.java
@@ -1,5 +1,0 @@
-package com.gamenawa.eeeeweeee.games;
-
-public interface IGameScoreSearcher {
-    GameScore getGameScoreByTitle(String title);
-}
