+++ src/main/java/com/gamenawa/eeeeweeee/global/util/json/GsonParser.java
@@ -1,4 +1,4 @@
-package com.gamenawa.eeeeweeee.util;
+package com.gamenawa.eeeeweeee.global.util.json;
 
 import com.google.gson.Gson;
 import org.springframework.stereotype.Component;
